--- .vite/deps_temp_48fc2343/package.json
@@ -1,3 +1,0 @@
-{
-  "type": "module"
-}
