+++ .vscode/extensions.json
@@ -1,3 +1,3 @@
 {
-  "recommendations": ["Vue.volar"]
+  "recommendations": ["Vue.volar", "Vue.vscode-typescript-vue-plugin"]
 }
