+++ components.d.ts
@@ -7,21 +7,14 @@ export {}
 /* prettier-ignore */
 declare module 'vue' {
   export interface GlobalComponents {
-    ContextMenu: typeof import('./src/components/Map/ContextMenu.vue')['default']
+    ContextMenu: typeof import('./src/client/components/Map/ContextMenu.vue')['default']
     ElBadge: typeof import('element-plus/es')['ElBadge']
     ElButton: typeof import('element-plus/es')['ElButton']
-    GameMap: typeof import('./src/components/Map/GameMap.vue')['default']
-    HelloWorld: typeof import('./src/components/HelloWorld.vue')['default']
-    IconCommunity: typeof import('./src/components/icons/IconCommunity.vue')['default']
-    IconDocumentation: typeof import('./src/components/icons/IconDocumentation.vue')['default']
-    IconEcosystem: typeof import('./src/components/icons/IconEcosystem.vue')['default']
-    IconSupport: typeof import('./src/components/icons/IconSupport.vue')['default']
-    IconTooling: typeof import('./src/components/icons/IconTooling.vue')['default']
-    Markdown: typeof import('./src/components/Model/Markdown.vue')['default']
-    MarkerLayer: typeof import('./src/components/Map/MarkerLayer.vue')['default']
+    GameMap: typeof import('./src/client/components/Map/GameMap.vue')['default']
+    HelloWorld: typeof import('./src/client/components/HelloWorld.vue')['default']
+    Markdown: typeof import('./src/client/components/Model/Markdown.vue')['default']
+    MarkerLayer: typeof import('./src/client/components/Map/MarkerLayer.vue')['default']
     RouterLink: typeof import('vue-router')['RouterLink']
     RouterView: typeof import('vue-router')['RouterView']
-    TheWelcome: typeof import('./src/components/TheWelcome.vue')['default']
-    WelcomeItem: typeof import('./src/components/WelcomeItem.vue')['default']
   }
 }
