+++ package.json
@@ -1,7 +1,7 @@
 {
-  "name": "game_map_new",
-  "version": "0.0.0",
+  "name": "map",
   "private": true,
+  "version": "0.0.0",
   "type": "module",
   "scripts": {
     "dev": "vite",
@@ -9,37 +9,46 @@
     "preview": "vite preview",
     "test:unit": "vitest",
     "build-only": "vite build",
-    "type-check": "vue-tsc --build --force"
+    "type-check": "vue-tsc --build --force",
+    "server:dev": "nodemon -w src/server -x tsx src/server/main.ts",
+    "server:start": "cross-env NODE_ENV=production tsx src/server/main.ts",
+    "server:build": "vite build"
   },
   "dependencies": {
     "@mdi/font": "^7.4.47",
     "axios": "^1.7.7",
     "canvaskit-wasm": "^0.39.1",
     "element-plus": "^2.8.6",
+    "express": "^4.21.1",
     "github-markdown-css": "^5.7.0",
     "marked": "^14.1.3",
     "ol": "^10.2.1",
     "pinia": "^2.2.4",
+    "vite-express": "^0.19.0",
     "vue": "^3.5.12",
     "vue-router": "^4.4.5",
     "vuetify": "^3.7.3"
   },
   "devDependencies": {
     "@tsconfig/node20": "^20.1.4",
+    "@types/express": "^4.17.21",
     "@types/jsdom": "^21.1.7",
-    "@types/node": "^20.17.0",
-    "@vitejs/plugin-vue": "^5.1.4",
+    "@types/node": "^20.9.3",
+    "@vitejs/plugin-vue": "^4.5.0",
     "@vue/test-utils": "^2.4.6",
     "@vue/tsconfig": "^0.5.1",
     "jsdom": "^25.0.1",
     "less": "^4.2.0",
+    "nodemon": "^3.0.1",
     "npm-run-all2": "^7.0.1",
+    "tsx": "^4.19.2",
     "typescript": "~5.6.0",
     "unplugin-auto-import": "^0.18.3",
+    "unplugin-icons": "^0.20.1",
     "unplugin-vue-components": "^0.27.4",
-    "vite": "^5.4.10",
+    "vite": "^5.0.2",
     "vite-plugin-vue-devtools": "^7.5.3",
     "vitest": "^2.1.3",
-    "vue-tsc": "^2.1.6"
+    "vue-tsc": "^1.8.22"
   }
 }
