--- src/App.vue
@@ -1,11 +1,0 @@
-<script setup lang="ts">
-import { RouterLink, RouterView } from 'vue-router'
-</script>
-
-<template>
-    <v-layout>
-        <RouterView />
-    </v-layout>
-</template>
-
-<style scoped></style>
