+++ src/client/App.vue
@@ -1,0 +1,11 @@
+<script setup lang="ts">
+import { RouterLink, RouterView } from 'vue-router'
+</script>
+
+<template>
+    <v-layout>
+        <RouterView />
+    </v-layout>
+</template>
+
+<style scoped></style>
