+++ src/client/plugins/vuetify.ts
@@ -1,0 +1,13 @@
+import 'vuetify/styles'
+import '@mdi/font/css/materialdesignicons.css'
+import { createVuetify } from 'vuetify'
+import * as components from 'vuetify/components'
+import * as directives from 'vuetify/directives'
+
+export const vuetify = createVuetify({
+    components,
+    directives,
+    theme: {
+        defaultTheme: 'dark'
+    }
+})
