+++ src/client/router/index.ts
@@ -1,0 +1,15 @@
+import { createRouter, createWebHistory } from 'vue-router'
+
+const router = createRouter({
+    history: createWebHistory(import.meta.env.BASE_URL),
+    routes: [
+        {
+            path: '/',
+            name: 'home',
+            component: () => import('../views/HomeView.vue')
+        },
+
+    ]
+})
+
+export default router
