--- src/main.ts
@@ -1,15 +1,0 @@
-
-import { createApp } from 'vue'
-import { createPinia } from 'pinia'
-
-import App from './App.vue'
-import router from './router'
-import { vuetify } from '@/Plugins/vuetify'
-
-const app = createApp(App)
-
-app.use(createPinia())
-app.use(router)
-app.use(vuetify)
-
-app.mount('#app')
