--- src/views/HomeView.vue
@@ -1,13 +1,0 @@
-<script lang='ts' setup>
-
-</script>
-<template>
-    <GameMap></GameMap>
-</template>
-<script lang='ts'>
-
-export default {
-    name: 'HomeView',
-}
-</script>
-<style lang='less' scoped></style>
