+++ vite.config.ts
@@ -7,23 +7,43 @@ import vueDevTools from 'vite-plugin-vue-devtools'
 import AutoImport from 'unplugin-auto-import/vite'
 import Components from 'unplugin-vue-components/vite'
 import { ElementPlusResolver } from 'unplugin-vue-components/resolvers'
+import IconsResolver from 'unplugin-icons/resolver'
+import Icons from 'unplugin-icons/vite'
 
-
-// https://vite.dev/config/
+// https://vite.dev/config/ 
 export default defineConfig({
     plugins: [
         vue(),
         vueDevTools(),
         AutoImport({
-            resolvers: [ElementPlusResolver()],
+            imports: ['vue'],
+            resolvers: [
+                ElementPlusResolver(),
+                // 自动导入图标组件
+                IconsResolver({
+                    prefix: 'Icon',
+                }),
+            ],
+
         }),
         Components({
-            resolvers: [ElementPlusResolver()],
+            dirs: ['src/client/components'],
+            resolvers: [
+                ElementPlusResolver(),
+                // 自动注册图标组件
+                IconsResolver({
+                    enabledCollections: ['ep'],
+                }),
+            ],
+        }),
+        Icons({
+            autoInstall: true,
         }),
     ],
     resolve: {
         alias: {
-            '@': fileURLToPath(new URL('./src', import.meta.url))
+            '@': fileURLToPath(new URL('./src/client', import.meta.url)),
+            '@srv': fileURLToPath(new URL('./src/server', import.meta.url)),
         }
     }
 })
